requirements.txt
google-generativeai
python-telegram-bot